# feat: enhance wafer CLI tool with mochi repository patterns