--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,10 @@
+module wafer
+
+go 1.24
+
+require (
+	github.com/alexflint/go-arg v1.5.1
+	github.com/google/uuid v1.6.0
+)
+
+require github.com/alexflint/go-scalar v1.2.0 // indirect