… .goreleaser.yaml
… BENCHMARK.md
… CHANGELOG.md
… CONTRIBUTING.md
… README.md
… ROADMAP.md
… SPEC.md
… VERSION
… bench/results.txt
… docs/features/embedding.md
… docs/guides/installation.md
… docs/guides/usage.md
… go.sum
… internal/ingest/embedder.go
… internal/ingest/processor.go
… tests/benchmark_test.go
… tests/fixtures/sample.txt
… tests/golden/multiline.jsonl
… tests/golden/simple.jsonl
… tests/golden_test.go
… tests/integration_test.go
… tests/unit_test.go