## 🚀 Major Enhancement: Wafer CLI Tool with Mochi Patterns

This PR implements comprehensive enhancements to the wafer CLI tool by adopting proven patterns and practices from the [mochilang/mochi](https://github.com/mochilang/mochi) repository.

### ✨ Key Enhancements

#### 🧪 Golden File Testing Framework
- **Regression testing system** with deterministic output validation
- **Output normalization** for stable comparisons (timestamps, UUIDs, paths)
- **Golden file update mechanism** with `--update` flag support
- **Test fixtures** in `tests/golden/` with expected outputs

#### ⚙️ Enhanced Build System
- **Sophisticated Makefile** with mochi's patterns and best practices
- **Colored output with emojis** for better developer experience
- **Parallel execution** for build targets (`make build-all`)
- **Comprehensive help system** with categorized targets
- **Tool detection and fallbacks** (golangci-lint, benchcmp)
- **Development environment automation** (`dev-setup`, `dev-check`)

#### 📊 Performance Benchmarking
- **Complete benchmark suite** for all core components
- **Baseline comparison system** (`bench-baseline`, `bench-compare`)
- **CPU profiling support** (`bench-profile`)
- **Memory allocation tracking** with detailed metrics
- **Performance regression detection** capabilities

#### 📚 Comprehensive Documentation
- **Enhanced README.md** with mochi's structure and quality standards
- **Technical ROADMAP.md** with detailed future planning
- **Performance BENCHMARK.md** with detailed metrics and analysis
- **Professional CONTRIBUTING.md** with clear guidelines
- **Improved SPEC.md** with formal specification structure

#### 🐳 Docker & Release Enhancements
- **Multi-stage Dockerfile** for optimized container builds
- **Docker Compose examples** for easy deployment
- **Cross-platform container support** (linux/amd64, linux/arm64)
- **GoReleaser configuration** for automated releases

### 📋 Files Added/Modified

#### Core Implementation
- `cmd/wafer/main.go` - CLI entrypoint with structured argument parsing
- `internal/config/config.go` - Configuration management
- `internal/ingest/` - Core processing logic (chunker, embedder, processor, writer)
- `go.mod`, `go.sum` - Go module definition and dependencies

#### Testing Framework
- `golden/golden.go` - Golden file testing framework
- `tests/golden_test.go` - Golden file regression tests
- `tests/benchmark_test.go` - Comprehensive benchmark suite
- `tests/unit_test.go` - Enhanced unit tests
- `tests/integration_test.go` - End-to-end integration tests
- `tests/fixtures/` - Test data files
- `tests/golden/` - Golden file test fixtures

#### Build & Release
- `Makefile` - Sophisticated build system with 25+ targets
- `.goreleaser.yaml` - Cross-platform release automation
- `Dockerfile` - Multi-stage container builds
- `VERSION` - Version management

#### Documentation
- `README.md` - Comprehensive project overview
- `ROADMAP.md` - Detailed future planning
- `BENCHMARK.md` - Performance metrics and analysis
- `CONTRIBUTING.md` - Professional contribution guidelines
- `SPEC.md` - Enhanced technical specification
- `CHANGELOG.md` - Detailed version history
- `docs/` - Structured documentation

### 🎯 Makefile Targets

#### Build Targets ⚙️
```bash
make build          # Build binary for current platform
make build-all      # Build for all platforms (parallel)
make release        # Build release binaries
```

#### Testing Targets 🧪
```bash
make test           # Run all tests with coverage
make test-golden    # Run golden file tests
make update-golden  # Update golden files
make bench          # Run benchmarks
make bench-profile  # Run with CPU profiling
```

#### Quality Targets 🔍
```bash
make lint           # Run static analysis
make fmt            # Format source code
make dev-setup      # Set up development environment
make dev-check      # Check development environment
```

#### Docker Targets 🐳
```bash
make docker-build   # Build Docker image
make docker-test    # Test Docker image
make docker-run     # Run with test fixtures
```

### 📊 Performance Benchmarks

```
BenchmarkChunker/Small-4            102585    11810 ns/op
BenchmarkChunker/Medium-4            12961    92358 ns/op
BenchmarkChunker/Large-4              1111  1073308 ns/op
BenchmarkChunkerMemory-4             12888    93891 ns/op   48656 B/op   1020 allocs/op
BenchmarkJSONLWriting-4              10000   101696 ns/op   13855 B/op      5 allocs/op
BenchmarkTextProcessingPipeline-4    2706   435486 ns/op  109296 B/op   1048 allocs/op
```

### 🧪 Testing Strategy

- **Unit Tests**: Individual component testing with mocks
- **Integration Tests**: End-to-end workflow testing
- **Golden File Tests**: Regression testing with deterministic output
- **Benchmark Tests**: Performance monitoring and regression detection
- **Docker Tests**: Container functionality validation

### 🌟 Quality Improvements

- **Code Coverage**: Comprehensive test coverage across all components
- **Performance Monitoring**: Baseline comparison and regression detection
- **Documentation**: Professional-grade documentation following industry standards
- **Developer Experience**: Enhanced tooling with colored output and automation
- **Build Automation**: Sophisticated build system with parallel execution

### 🎯 Following Mochi's Best Practices

✅ **Build System Enhancement** - Implemented mochi's sophisticated Makefile patterns
✅ **Golden File Testing** - Adopted mochi's regression testing framework
✅ **Documentation Standards** - Enhanced with mochi's structure and quality
✅ **Performance Monitoring** - Implemented comprehensive benchmark suite
✅ **Professional Organization** - Following mochi's proven project structure

### 🚀 Ready for Production

This enhancement brings wafer to **production-ready standards** with:

- ✅ **Comprehensive testing** with multiple test types
- ✅ **Professional build system** with automation
- ✅ **Performance monitoring** with regression detection
- ✅ **Docker support** for containerized deployment
- ✅ **Professional documentation** for users and contributors
- ✅ **Enhanced developer experience** with colored output and tooling

### 🔍 Review Checklist

- [x] All tests pass (`make test`)
- [x] Code is formatted (`make fmt`)
- [x] Linting passes (`make lint`)
- [x] Documentation is comprehensive
- [x] Benchmarks establish baseline
- [x] Docker builds successfully
- [x] Golden tests validate output format

---

**This PR transforms wafer from a basic CLI tool into a professional, production-ready project following the proven patterns established by the mochilang/mochi repository.**

**Note**: GitHub Actions workflows will be added in a follow-up commit due to OAuth scope limitations.

---
Pull Request opened by [Augment Code](https://www.augmentcode.com/) with guidance from the PR author